Better handling of failure to open dump file.
On an early panic, dump_format would be called
before meaningful values existed for %n, etc.
This could result in chmod() being called on a
parent directory such as userdata, resulting in
a complete denial of service. Also, dump file
write was being attempted even if file failed
to open.